# Conflicts:
#	apps/accounts/schemas/business.py
#	apps/accounts/schemas/common.py
#	apps/accounts/tests/test_business.py
#	apps/accounts/views/business.py
#	apps/core/models.py
#	apps/core/scheduler.py
#	apps/core/tasks.py
#	apps/jobs/models.py
#	apps/settings/schemas.py
#	apps/settings/tests/test_views.py
#	helpers/email/auth.py

+++ apps/accounts/models.py
@@ -9,7 +9,7 @@ from uuid import UUID
 
 import jwt
 from accounts.enums import UserType, AuthType, GenderType, BusinessUserRoleType, NoticePeriodType, Months, Days, \
-    BusinessSize, BusinessUserStatusType, CaseReasonType
+    BusinessSize, BusinessUserStatusType, CaseReasonType, AdminRoleType
 from core.enums import SalaryType
 from core.models import BaseModel, State, City
 from dateutil.relativedelta import relativedelta
@@ -17,7 +17,7 @@ from django.contrib.auth.hashers import check_password, make_password
 from django.contrib.auth.models import AbstractUser, BaseUserManager
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.db.models import Q, Count, F, Value, Avg, IntegerField, Exists, OuterRef, Sum, When, Case
+from django.db.models import Q, Count, F, Value, Avg, IntegerField, Exists, OuterRef, When, Case, Sum
 from django.db.models.functions import Concat, Cast, Round
 from django.db.models.signals import pre_save
 from django.utils import timezone
@@ -163,6 +163,12 @@ class User(AbstractUser, BaseModel):
     def delete_account(self):
         self.hard_delete()
         return
+
+
+class AdminUser(BaseModel):
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    role = models.CharField(max_length=50, choices=AdminRoleType.choices())
+
 
 
 class Country(BaseModel):
@@ -376,7 +382,7 @@ class Talent(BaseModel):
 
         # Query Job table with optimized prefetch and select_related
         if business:
-            job_matching_query &= Q(created_by__business=business)
+            job_matching_query &= Q(created_by__business=business, created_by__business__paused=False)
         jobs = (Job.objects.prefetch_related("requiredattribute", "additional_languages", "skills", 'jobpost', 'availableday')
                 .filter(job_matching_query)
                 .only("id")
@@ -436,7 +442,7 @@ class Talent(BaseModel):
         is_saved = Exists(SavedJob.objects.filter(job_post__id=OuterRef("id"), talent=self))
         queryset = JobPost.objects.select_related(
             "job", "country", "job__role", "job__created_by__business"
-        ).annotate(is_saved=is_saved).filter(is_saved=True)
+        ).annotate(is_saved=is_saved).filter(is_saved=True, job__created_by__business__paused=False)
         queryset = add_job_post_annotations(queryset, self)
         return queryset
 
@@ -447,7 +453,7 @@ class Talent(BaseModel):
         is_applied = Exists(JobApplication.objects.filter(job_post__id=OuterRef("id"), applicant=self))
         queryset = JobPost.objects.select_related(
             "job", "country", "job__role", "job__created_by__business"
-        ).annotate(is_applied=is_applied).filter(is_applied=True)
+        ).annotate(is_applied=is_applied).filter(is_applied=True, job__created_by__business__paused=False)
         queryset = add_job_post_annotations(queryset, self)
         return queryset
 
@@ -522,7 +528,7 @@ class Talent(BaseModel):
             notifications = notifications.exclude(entity__in=excludes)
         return notifications.order_by("-id")
 
-    def delete_account(self):
+    def delete_account(self, banned=False):
         self.savedjob_set.all().hard_delete()
         self.education_set.all().hard_delete()
         self.experience_set.all().hard_delete()
@@ -551,7 +557,12 @@ class Talent(BaseModel):
             self.cv.delete()
         self.cv = None
         self.save()
-        self.delete()
+        if banned is False:
+            self.delete()
+        else:
+            self.jobapplication_set.all().hard_delete()
+            self.jobapplicationwithdrawal_set.all().hard_delete()
+            self.hard_delete()
 
     def is_profile_completed(self, raise_exception=False):
         # Check education
@@ -659,6 +670,7 @@ class Business(BaseModel):
     facebook = models.URLField(null=True)
     twitter_x = models.URLField(null=True)
     industry = models.ForeignKey(BusinessIndustry, null=True, on_delete=models.SET_NULL)
+    paused = models.BooleanField(default=False)
 
     def __str__(self):
         return f"{self.name}"
@@ -1149,8 +1161,11 @@ class Business(BaseModel):
             query["status"] = status
         return JobPost.objects.select_related("job__created_by__business").filter(**query)
 
+    def reg_date(self):
+        return self.created_at.date()
+
 class BusinessClient(BaseModel):
-    business = models.ForeignKey("accounts.Business", on_delete=models.CASCADE)
+    business = models.ForeignKey("accounts.Business", on_delete=models.CASCADE, related_name="clients")
     name = models.CharField(max_length=225)
 
 class VerificationCode(BaseModel):
@@ -1227,8 +1242,9 @@ class BusinessUser(BaseModel):
         return Notification.objects.none()
 
     def delete_account(self):
-        self.user.hard_delete()
+        user = self.user
         self.hard_delete()
+        user.delete_account()
         return
 
     @property
@@ -1250,6 +1266,10 @@ class BusinessUser(BaseModel):
             return data
         except jwt.DecodeError:
             return None
+
+
+    def date_joined(self):
+        return self.created_at.date()
 
 
 class Education(BaseModel):
@@ -1341,3 +1361,7 @@ class TalentFilter(BaseModel):
     skills = models.ManyToManyField("accounts.Skill", related_name="skills_talent_filter")
     business_models = models.ManyToManyField("jobs.BusinessModel", blank=True, related_name="businessmodels_talent_filter")
     completed_profiles = models.BooleanField(default=True)
+
+class BannedAccount(BaseModel):
+    email = models.EmailField()
+    account_type = models.CharField(choices=UserType.choices())
